moved instance to own file. moved functions based on comments in pr
