+++ libraries/platform/src/platform.h
@@ -9,41 +9,15 @@
 #ifndef hifi_Platform_h
 #define hifi_Platform_h
 
+#include "platformInstance.h"
 #include <vector>
 #include <nlohmann/json.hpp>
 
 namespace platform {
     using json = nlohmann::json;
     
-class Instance {
-public:
-    bool virtual enumeratePlatform() = 0;
-
-    int getNumCPU() { return (int)_cpu.size(); }
-    json getCPU(int index);
-
-    int getNumGPU() { return (int)_gpu.size(); }
-    json getGPU(int index);
-
-    int getNumMemory() { return (int)_memory.size(); }
-    json getMemory(int index);
-
-    int getNumDisplay() { return (int)_display.size(); }
-    json getDisplay(int index);
-
-    virtual ~Instance();
-
-protected:
-    std::vector<json>  _cpu;
-    std::vector<json>  _memory;
-    std::vector<json>  _gpu;
-    std::vector<json>  _display;
-};
-
-//Platform level functions
 void create();
 void destroy();
-
 bool enumeratePlatform();
 
 int getNumCPU();
